Add getScreenSize fun , AnalyticsScreen Added LineChart And Fix LineChart

+++ composeApp/src/commonMain/kotlin/feature/analytics/presentation/model/ValueLabel.kt
@@ -1,5 +1,5 @@
 package feature.analytics.presentation.model
 
 data class ValueLabel(
-    val value: Float
+    val value: Long
 )
